backend/stripe-payment/requirements.txt
psycopg2-binary==2.9.9
stripe==7.8.0
